pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.poetry]
name = "pre-commit-hooks"
version = "0.0.0"
description = "Pre-Commit Hooks"
authors = [
    "Chris Laco <claco@chrislaco.com>",
]
packages = [
    { include = "pre_commit_hooks" }
]

[tool.poetry.dependencies]
gitpython = "*"
toml = "*"

[tool.poetry.scripts]
check-authors = "pre_commit_hooks.check_authors:main"
check-changelog = "pre_commit_hooks.check_changelog:main"

[tool.pre_commit_hooks.check_authors]

[tool.pre_commit_hooks.check_changelog]
